requirements.txt
# Core computational libraries
numpy>=1.21.0
scipy>=1.7.0
numba>=0.54.0

# Data handling and configuration
pandas>=1.3.0
pyyaml>=5.4.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Logging and utilities
colorlog>=6.6.0
tqdm>=4.62.0

# Optional for performance
h5py>=3.0.0  # For saving large datasets